@@ -15,7 +15,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.0.1+3
 
 environment:
   sdk: ">=2.16.1 <3.0.0"
@@ -51,8 +51,8 @@ dependencies:
   get_storage: ^2.0.2
   google_fonts: ^2.3.1
   google_maps_flutter: ^2.1.2
-  image_cropper: ^1.4.0
-  image_picker: any
+  image_cropper: ^3.0.0
+  image_picker: ^0.8.5+3
   internet_connection_checker: ^0.0.1+1
   intl: ^0.17.0
   logger: ^1.1.0
@@ -65,9 +65,15 @@ dependencies:
   geolocator: ^8.2.1
   carousel_slider: ^4.1.1
   pinch_zoom: ^1.0.0
-  onesignal_flutter: ^3.3.2
+  onesignal_flutter: ^3.4.0
+  package_info_plus: any
   persistent_bottom_nav_bar: ^4.0.2
-  flutter_colorpicker: ^1.0.3
+  dotted_line: ^3.1.0
+  webview_flutter: ^3.0.4
+  flutter_rating_bar: ^4.0.1
+  flutter_animation_progress_bar: ^2.2.1
+  launch_review: ^1.0.1
+
 
   
 
