@@ -2,13 +2,17 @@ import 'dart:convert';
 
 import 'package:animated_bottom_navigation_bar/animated_bottom_navigation_bar.dart';
 import 'package:flutter/material.dart';
+import 'package:market_vendor_app/activities/home_page/tab/my_order.dart';
+import 'package:market_vendor_app/utils/notification_receiver.dart';
 
 import '../../theme/app_colors.dart';
 import '../../theme/styles.dart';
 import '../../utils/globals.dart' as globals;
+import '../../utils/notification_show.dart';
 import '../../utils/shared_preferences.dart';
 import '../add_services/add_product_and_service_view.dart';
 
+import 'drawer_screen/rating_screen.dart';
 import 'drawer_view.dart';
 import 'model/profile_model.dart';
 import 'tab/bell_screen.dart';
@@ -22,12 +26,13 @@ class TabScreen extends StatefulWidget {
   _TabScreenState createState() => _TabScreenState();
 }
 
-class _TabScreenState extends State<TabScreen> {
+class _TabScreenState extends State<TabScreen>
+    implements NotificationInterface {
   final iconList = <IconData>[
     Icons.home,
     Icons.shopping_bag_outlined,
-    Icons.notifications_none,
-    Icons.brightness_7,
+    Icons.star,
+    Icons.account_circle,
   ];
   var profileData;
   ProfileModel model = ProfileModel();
@@ -36,6 +41,11 @@ class _TabScreenState extends State<TabScreen> {
   @override
   void initState() {
     super.initState();
+    NotificationShow.initPlatformState(this);
+    getData();
+  }
+
+  getData() {
     SharedPref.getProfileData().then((value) {
       profileData = json.decode(value);
       model = ProfileModel.fromJson(profileData);
@@ -49,19 +59,56 @@ class _TabScreenState extends State<TabScreen> {
   var _bottomNavIndex = 0;
   AnimationController? _animationController;
   Animation<double>? animation;
+  String totalorder = "";
+  String totalRevenue = "";
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      drawer: DrawerViewScreen(),
+      drawer: DrawerViewScreen(
+        model: model,
+        totalOrder: totalorder,
+        totalRevenue: totalRevenue,
+        callBack: (v) {
+          _bottomNavIndex = v;
+          setState(() {});
+        },
+      ),
       body: Container(
           child: _bottomNavIndex == 0
-              ? HomeScreen()
+              ? HomeScreen(
+                  totalOrders: (v) {
+                    totalorder = v;
+                    setState(() {});
+                  },
+                  totalRevenue: (v) {
+                    totalRevenue = v;
+                    setState(() {});
+                  },
+                )
               : _bottomNavIndex == 1
-                  ? ShopingCartScreen()
+                  ? MyOrderScreen(
+                      callBack: (v) {
+                        _bottomNavIndex = v;
+                        getData();
+                        setState(() {});
+                      },
+                    )
                   : _bottomNavIndex == 2
-                      ? BellScreen()
+                      ? RatingScreen(
+                          callBack: (v) {
+                            _bottomNavIndex = v;
+                            getData();
+                            setState(() {});
+                          },
+                        )
                       : _bottomNavIndex == 3
-                          ? SettingScreen()
+                          ? MyAccountSettingScreen(
+                              callBack: (v) {
+                                _bottomNavIndex = v;
+                                getData();
+                                setState(() {});
+                              },
+                            )
                           : AddProductAndServiceScreen(
                               isProductScreen: false,
                             )),
@@ -74,13 +121,11 @@ class _TabScreenState extends State<TabScreen> {
           color: Colors.white,
         ),
         onPressed: () {
-          if (model.data!.businesses!.isVerified.toString() == "Yes") {
-            setState(() {
-              _bottomNavIndex = -1;
-            });
-            _animationController!.reset();
-            _animationController!.forward();
-          } else {}
+          setState(() {
+            _bottomNavIndex = -1;
+          });
+          _animationController!.reset();
+          _animationController!.forward();
         },
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
@@ -111,5 +156,15 @@ class _TabScreenState extends State<TabScreen> {
         onTap: (index) => setState(() => _bottomNavIndex = index),
       ),
     );
+  }
+
+  @override
+  void onClick(id, type, requestId) {
+    // TODO: implement onClick
+  }
+
+  @override
+  void onMessageReceived(id, type) {
+    // TODO: implement onMessageReceived
   }
 }
