@@ -1,81 +1,80 @@
+import 'dart:convert';
+
+import 'package:market_vendor_app/utils/notification_receiver.dart';
+import 'package:market_vendor_app/utils/shared_preferences.dart';
+import 'package:market_vendor_app/utils/strings/app_constants.dart';
 import 'package:onesignal_flutter/onesignal_flutter.dart';
 
 class NotificationShow {
-  static Future<void> initPlatformState() async {
-    OneSignal.shared.setLogLevel(OSLogLevel.verbose, OSLogLevel.none);
+  // static initPlatformState(NotificationInterface callBack) async {
+  //   OneSignal.shared.init(AppConstants.oneSingnalAppId, iOSSettings: {
+  //     OSiOSSettings.autoPrompt: true,
+  //     OSiOSSettings.inAppLaunchUrl: true
+  //   });
+  //   OneSignal.shared
+  //       .setInFocusDisplayType(OSNotificationDisplayType.notification);
+  //   var status = await OneSignal.shared.getPermissionSubscriptionState();
+  //   var playerId = status.subscriptionStatus.userId;
+  //   print("One Signal Player ID == > $playerId ");
+  //   SharedPref.savePlayerId(playerId);
 
-    OneSignal.shared.setRequiresUserPrivacyConsent(true);
+  //   OneSignal.shared.setNotificationReceivedHandler((notification) {
+  //     // int ID = notification.androidNotificationId;
+  //     print("addinal data ${notification.payload.additionalData}");
+
+  //     var data = json.encode(notification.payload.additionalData);
+  //     Map p = jsonDecode(data);
+  //     // SharedPref.setOrderID(p['order_id'].toString());
+  //     // SharedPref.setOrderType(p['type']);
+  //     // SharedPref.setRequestOrderId(p['request_id'].toString());
+  //     callBack.onMessageReceived(p['order_id'], p['type']);
+  //   });
+  //   OneSignal.shared
+  //       .setNotificationOpenedHandler((OSNotificationOpenedResult result) {
+  //     // print("result typedsfgfgdfgdgdgdgdgdg");
+  //     print("addinal data ${result.notification.payload.additionalData}");
+
+  //     var data = json.encode(result.notification.payload.additionalData);
+  //     Map p = jsonDecode(data);
+  //     // print("result type==========" +
+  //     //     result.notification.payload.rawPayload!.toString());
+  //     SharedPref.setOrderID(p['order_id'].toString());
+  //     SharedPref.setOrderType(p['type']);
+  //     SharedPref.setRequestOrderId(p['request_id'].toString());
+  //     callBack.onClick(p['order_id'], p['type'], p['request_id']);
+  //   });
+  //   OneSignal.shared
+  //       .setInAppMessageClickedHandler((OSInAppMessageAction action) {});
+  // }
+
+  static initPlatformState(NotificationInterface callBack) async {
+    await OneSignal.shared.setAppId(AppConstants.oneSingnalAppId);
+    final status = await OneSignal.shared.getDeviceState();
+    String? osUserID = status?.userId;
+    print("One Signal Player ID == > $osUserID ");
+    SharedPref.savePlayerId(osUserID!);
+    await OneSignal.shared.promptUserForPushNotificationPermission(
+      fallbackToSettings: true,
+    );
 
     OneSignal.shared
         .setNotificationOpenedHandler((OSNotificationOpenedResult result) {
       print('NOTIFICATION OPENED HANDLER CALLED WITH: ${result}');
-      // this.setState(() {
-      //   _debugLabelString =
-      //       "Opened notification: \n${result.notification.jsonRepresentation().replaceAll("\\n", "\n")}";
-      // });
+      print("addinal data ${result.notification.additionalData}");
+
+      var data = json.encode(result.notification.additionalData);
+      Map p = jsonDecode(data);
+      // print("result type==========" +
+      //     result.notification.payload.rawPayload!.toString());
+      SharedPref.setOrderID(p['order_id'].toString());
+      SharedPref.setOrderType(p['type']);
+      SharedPref.setRequestOrderId(p['request_id'].toString());
+      callBack.onClick(p['order_id'], p['type'], p['request_id']);
     });
 
     OneSignal.shared.setNotificationWillShowInForegroundHandler(
         (OSNotificationReceivedEvent event) {
       print('FOREGROUND HANDLER CALLED WITH: ${event}');
-
-      /// Display Notification, send null to not display
-      event.complete(null);
-
-      // this.setState(() {
-      //   _debugLabelString =
-      //       "Notification received in foreground notification: \n${event.notification.jsonRepresentation().replaceAll("\\n", "\n")}";
-      // });
     });
-
-    OneSignal.shared
-        .setInAppMessageClickedHandler((OSInAppMessageAction action) {
-      // this.setState(() {
-      //   _debugLabelString =
-      //       "In App Message Clicked: \n${action.jsonRepresentation().replaceAll("\\n", "\n")}";
-      // });
-    });
-
-    OneSignal.shared
-        .setSubscriptionObserver((OSSubscriptionStateChanges changes) {
-      print("SUBSCRIPTION STATE CHANGED: ${changes.jsonRepresentation()}");
-    });
-
-    OneSignal.shared.setPermissionObserver((OSPermissionStateChanges changes) {
-      print("PERMISSION STATE CHANGED: ${changes.jsonRepresentation()}");
-    });
-
-    OneSignal.shared.setEmailSubscriptionObserver(
-        (OSEmailSubscriptionStateChanges changes) {
-      print("EMAIL SUBSCRIPTION STATE CHANGED ${changes.jsonRepresentation()}");
-    });
-
-    OneSignal.shared
-        .setSMSSubscriptionObserver((OSSMSSubscriptionStateChanges changes) {
-      print("SMS SUBSCRIPTION STATE CHANGED ${changes.jsonRepresentation()}");
-    });
-
-    // NOTE: Replace with your own app ID from https://www.onesignal.com
-    await OneSignal.shared.setAppId("c61cc644-eaf4-488c-83a3-2417fc9fe170");
-
-    // iOS-only method to open launch URLs in Safari when set to false
-    OneSignal.shared.setLaunchURLsInApp(false);
-    var deviceState = await OneSignal.shared.getDeviceState();
-
-    if (deviceState == null || deviceState.userId == null) return;
-
-    var playerId = deviceState.userId!;
-
-    OneSignal.shared.disablePush(false);
-  }
-
-  static void _playerId() async {
-    var deviceState = await OneSignal.shared.getDeviceState();
-
-    if (deviceState == null || deviceState.userId == null) return;
-
-    var playerId = deviceState.userId!;
-
-    print(playerId);
   }
 }
